+++ api_gateway/src/main/kotlin/org/thechance/api_gateway/endpoints/PermissionRoute.kt
@@ -1,8 +1,11 @@
 package org.thechance.api_gateway.endpoints
 
 import io.ktor.http.*
+import io.ktor.server.application.*
+import io.ktor.server.request.*
 import io.ktor.server.routing.*
 import org.koin.ktor.ext.inject
+import org.thechance.api_gateway.data.model.restaurant.RestaurantRequestPermission
 import org.thechance.api_gateway.endpoints.gateway.IRestaurantGateway
 import org.thechance.api_gateway.endpoints.utils.authenticateWithRole
 import org.thechance.api_gateway.endpoints.utils.extractLocalizationHeader
@@ -10,13 +13,20 @@ import org.thechance.api_gateway.endpoints.utils.respondWithResult
 import org.thechance.api_gateway.util.Role
 import java.util.*
 
-//TODO rename it to permission and move all permissions to it
-
-fun Route.dashboardRoutes() {
+fun Route.permissionRoutes() {
 
     val restaurantGateway: IRestaurantGateway by inject()
 
-    route("/dashboard") {
+    route("/permission") {
+
+        post("/restaurant") {
+            val requestedForm = call.receive<RestaurantRequestPermission>()
+            val (language, countryCode) = extractLocalizationHeader()
+            val result = restaurantGateway.createRequestPermission(
+                requestedForm, Locale(language, countryCode)
+            )
+            respondWithResult(HttpStatusCode.Created, result)
+        }
 
         authenticateWithRole(Role.DASHBOARD_ADMIN) {
             get("/restaurant-permission-request") {
