+++ api_gateway/src/main/kotlin/org/thechance/api_gateway/plugins/Routing.kt
@@ -4,7 +4,7 @@ import io.ktor.server.application.*
 import io.ktor.server.routing.*
 import org.thechance.api_gateway.data.security.TokenConfiguration
 import org.thechance.api_gateway.endpoints.cuisineRoute
-import org.thechance.api_gateway.endpoints.dashboardRoutes
+import org.thechance.api_gateway.endpoints.permissionRoutes
 import org.thechance.api_gateway.endpoints.restaurantRoutes
 import org.thechance.api_gateway.endpoints.taxiRoutes
 import org.thechance.api_gateway.endpoints.userRoutes
@@ -12,8 +12,9 @@ import org.thechance.api_gateway.endpoints.*
 
 fun Application.configureRouting(tokenConfiguration: TokenConfiguration) {
     routing {
-        userRoutes(tokenConfiguration)
-        dashboardRoutes()
+        authenticationRoutes(tokenConfiguration)
+        userRoutes()
+        permissionRoutes()
         restaurantRoutes()
         cuisineRoute()
         taxiRoutes()
